Refactor: Added static methods to Extensions class. Added left and right navigation to IP field.
The IP field can now navigate left and right. It's not perfect, but it works quite nice. The OnPropertyChanged() is called when certain keys are pressed. This isnt ideal, but the code still works to restrict the input of anything but numbers. The default IP and Port numbers can still not be set externally.